Init Commit
Set up board. Created Stats script with children for Magical Stats and
Physical Stats. Created a CameraController script that allows the player
to rotate the camera around the center of the board and zoom in using
their fingers.  Created PlayerController script to allow the player to
target their game pieces to perform move functions and attack functions.